moving context to its own file and moving events under Context
